// pc/src/api/fx1/data-contracts.ts
export interface IPage {
  /** @format int64 */
  current?: number;
  /** @format int64 */
  pages?: number;
  records?: _3[];
  searchCount?: boolean;
  /** @format int64 */
  size?: number;
  /** @format int64 */
  total?: number;
}

/** IPage«维保单位、企业单位、高地大化重点单位关联表(南宁维保模块)» */
export interface IPage2 {
  /** @format int64 */
  current?: number;
  /** @format int64 */
  pages?: number;
  records?: _4[];
  searchCount?: boolean;
  /** @format int64 */
  size?: number;
  /** @format int64 */
  total?: number;
}

/** 分页查询«维保关键词信息表(南宁维保模块)» */
export interface _ {
  /**
   * 当前页码
   * @format int32
   * @min 1
   * @exclusiveMin false
   */
  currentPage?: number;
  /**
   * 页数据条数
   * @format int32
   * @min 0
   * @exclusiveMin false
   * @max 100
   * @exclusiveMax false
   */
  pageSize?: number;
  /** 查询条件 */
  query?: _3;
}

/** 分页查询«维保单位、企业单位、高地大化重点单位关联表(南宁维保模块)» */
export interface _2 {
  /**
   * 当前页码
   * @format int32
   * @min 1
   * @exclusiveMin false
   */
  currentPage?: number;
  /**
   * 页数据条数
   * @format int32
   * @min 0
   * @exclusiveMin false
   * @max 100
   * @exclusiveMax false
   */
  pageSize?: number;
  /** 查询条件 */
  query?: _4;
}

/** 维保关键词信息表(南宁维保模块) */
export interface _3 {
  /** 创建人用户帐号 */
  cjryhzh: string;
  /**
   * 创建时间
   * @format date-time
   */
  cjsj?: string;
  /**
   *  风险等级, 1低风险, 2中风险, 3高风险
   * @format int32
   */
  fxdj: number;
  /** 关键词 */
  gjc: string;
  /** 主键ID */
  gjcid: string;
  /**
   * 记录状态0: 已删除  1: 可用
   * @format int32
   * @example 1
   */
  jlzt: number;
  /**
   *  是否自动发送短信, 0否, 1是
   * @format int32
   */
  sfzdfsdx: number;
  /** 修改人用户账号 */
  xgryhzh?: string;
  /** 数据归属业务系统来源 */
  ywxtly: string;
  /** 数据最后更新时间 */
  zhgxsj?: string;
  /**
   *  启用状态, 0未启用, 1启用
   * @format int32
   */
  zt: number;
}
// ...
export interface Boolean {
  /** 代码 */
  code?: string;
  /** 数据 */
  data?: boolean;
  /** 提示消息 */
  msg?: object;
}

/** 返回结果«维保关键词信息表(南宁维保模块)» */
export interface _6 {
  /** 代码 */
  code?: string;
  /** 数据 */
  data?: _3;
  /** 提示消息 */
  msg?: object;
}
// ...
export interface DeleteUsingDeleteParams {
  /** 主键 */
  id: string;
}

export interface KeyUsingGetParams {
  /** 主键 */
  id: string;
}